Rename default files
