+++ gems/aws-sdk-core/lib/aws-defaults.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require_relative 'aws-defaults/default-configuration'
+require_relative 'aws-defaults/default_configuration'
