+++ gems/aws-sdk-core/lib/aws-defaults/configuration_mode_defaults_resolver.rb
