+++ gems/aws-sdk-core/spec/aws-defaults/configuration_mode_defaults_resolver_spec.rb
